feat(kernel): compact metadata and expose capacity evidence
Generate syscall numbers, argument counts, aligned handler dispatch, trace names, and errno text from local declarative masters. Preserve syscall 23 as __sysctl, the slot-66 vfork dispatch, the slot-2 public vfork alias, conditional network behavior, and the void handler(void) ABI while pooling strings and narrowing retained tables.

Record process, inode, file, clist, and buffer occupancy at the native allocation seams, paint and sample both RP2040 u-area stack regions, and expose the versioned result through kern.capacity. Keep every fixed-table limit, USIZE, and the USB transmit ring unchanged pending Renode and board workload evidence.

Validated with warning-fatal PICO and PICO_UART builds, a clean distribution, the full RP2040 check tier, STM32 link checks, PIC32 touched-unit compilation, flash artifact construction, generated-file negative controls, ELF/disassembly inspection, and check-renode login, capacity, and deliberate MPU-fault assertions. Hardware access was not performed.

+++ sys/kern/tty.c
@@ -87,7 +87,7 @@ const char partab[] = {
     0007,0007,0007,0007,0007,0007,0007,0007
 };
 
-const int tthiwat[NSPEEDS] = {
+const uint16_t tthiwat[NSPEEDS] = {
     100,    /* 0 baud */
     100,    /* 50 baud */
     100,    /* 75 baud */
@@ -119,7 +119,7 @@ const int tthiwat[NSPEEDS] = {
     2000,   /* 4000 kbaud */
 };
 
-const int ttlowat[NSPEEDS] = {
+const uint8_t ttlowat[NSPEEDS] = {
     30,     /* 0 baud */
     30,     /* 50 baud */
     30,     /* 75 baud */
